Movimiento de funciones al archivo data.js
